requirements.txt
typer>=0.9.0
pynacl>=1.5.0 